+++ IteaIOSDip/Assets.xcassets/Filter/codding.imageset/Contents.json
@@ -2,17 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "icoWatchDay.png",
+      "filename" : "codding.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "icoWatchDay@2x.png",
+      "filename" : "codding@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "icoWatchDay@3x.png",
+      "filename" : "codding@3x.png",
       "scale" : "3x"
     }
   ],
