+++ IteaIOSDip/Assets.xcassets/Filter/frontend.imageset/Contents.json
@@ -2,17 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "icowatchNight.png",
+      "filename" : "frontend.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "icowatchNight-1.png",
+      "filename" : "frontend@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "icowatchNight-2.png",
+      "filename" : "frontend@3x.png",
       "scale" : "3x"
     }
   ],
