--- IteaIOSDip/Assets.xcassets/Filter/icowatchNight.dataset/Contents.json
@@ -1,11 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  },
-  "data" : [
-    {
-      "idiom" : "universal"
-    }
-  ]
-}
