+++ IteaIOSDip/Assets.xcassets/Filter/icowatchNight.imageset/Contents.json
@@ -2,17 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "iconfinder_-_Filter-Filters_3844475.png",
+      "filename" : "icowatchNight.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "iconfinder_-_Filter-Filters_3844475-1.png",
+      "filename" : "icowatchNight@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "iconfinder_-_Filter-Filters_3844475-2.png",
+      "filename" : "icowatchNight@3x.png",
       "scale" : "3x"
     }
   ],
