+++ IteaIOSDip/Models/Courses/CoursesFlowViewController+Extension.swift
@@ -8,19 +8,14 @@
 
 import UIKit
 
+//  MARK: - SET BACKGROUND -
 extension CoursesFlowViewController {
     
     func uiDesignCoursesFlow() {
-        
         let backgroundImage = UIImageView(frame: UIScreen.main.bounds)
         backgroundImage.image = UIImage(named: "bg")
         backgroundImage.contentMode = .scaleAspectFill
-//        let backgroundForImage = UIView(frame: UIScreen.main.bounds)
-//        backgroundForImage.backgroundColor = UIColor.red
-//        backgroundForImage.alpha = 0.8
         view.addSubview(backgroundImage)
-//        view.addSubview(backgroundForImage)
-//        self.view.sendSubviewToBack(backgroundForImage)
         self.view.sendSubviewToBack(backgroundImage)
     }
 }
